--- .eslintrc.js
@@ -1,59 +1,0 @@
-module.exports = {
-  env: {
-    es6: true,
-    browser: true,
-    node: true,
-  },
-  extends: [
-    'airbnb',
-    'plugin:react/recommended',
-    'plugin:@typescript-eslint/recommended',
-  ],
-  plugins: ['import', 'react', 'prettier'],
-  parserOptions: {
-    ecmaVersion: 6,
-    sourceType: 'module',
-    ecmaFeatures: {
-      jsx: true,
-    },
-  },
-  settings: {
-    'import/extensions': ['.js', '.jsx', '.ts', '.tsx'],
-    'import/parsers': {
-      '@typescript-eslint/parser': ['.ts', '.tsx'],
-    },
-    'import/resolver': {
-      node: {
-        extensions: ['.js', '.jsx', '.ts', '.tsx'],
-      },
-    },
-  },
-  rules: {
-    'react/jsx-filename-extension': [
-      1,
-      {
-        extensions: ['.tsx', '.ts'],
-      },
-    ],
-
-    'import/extensions': [
-      'error',
-      'ignorePackages',
-      {
-        js: 'never',
-        jsx: 'never',
-        ts: 'never',
-        tsx: 'never',
-      },
-    ],
-
-    'jsx-a11y/media-has-caption': 0,
-    'react/state-in-constructor': 0,
-    'react/prop-types': 0,
-    'implicit-arrow-linebreak': 0,
-    'object-curly-newline': 0,
-    'react/function-component-definition': 0,
-    'import/prefer-default-export': 0,
-    'operator-linebreak': 0,
-  },
-};
