+++ package.json
@@ -8,7 +8,7 @@
     }
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-app-rewired start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -45,7 +45,6 @@
     "@types/react-dom": "^18.0.10",
     "@typescript-eslint/eslint-plugin": "^5.50.0",
     "@typescript-eslint/parser": "^5.50.0",
-    "sass": "^1.57.1",
     "eslint": "^8.33.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.6.0",
@@ -55,6 +54,8 @@
     "eslint-plugin-react": "^7.32.2",
     "husky": "^4.2.5",
     "prettier": "^2.8.3",
+    "react-app-rewired": "^2.2.1",
+    "sass": "^1.57.1",
     "stylelint": "^14.16.1",
     "stylelint-config-htmlacademy": "^1.1.4",
     "stylelint-config-prettier": "^9.0.4",
