+++ src/helpers/data/en/tv-shows.ts
@@ -1,5 +1,5 @@
-import { IObjectOfStrings } from '../../types';
-import { getShowsCategories } from '../../utils/get-shows-categories';
+import { IObjectOfStrings } from '@types';
+import { getShowsCategories } from '@utils/get-shows-categories';
 import { CATEGORIES } from './dictionary';
 
 const { detective, fantastic, drama, comedy, medicine, final } =
