+++ src/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
+import { App } from '@components/App';
 import './index.scss';
-
-import { App } from './components/App';
 
 const container = document?.querySelector('.app');
