+++ src/pages/Play/PlayPage.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
-import { IPlayPage, IShow } from '../../helpers/types';
-import { DICTIONARY } from '../../helpers/data';
-import { Header } from '../../components/Header';
-import { GameProgress } from '../../components/GameProgress';
-import { Question } from '../../components/Question';
-import { AnswersList } from '../../components/AnswersList';
-import { ShowInfo } from '../../components/ShowInfo';
+import { IPlayPage, IShow } from '@types';
+import { DICTIONARY } from '@data';
+import { Header } from '@components/Header';
+import { GameProgress } from '@components/GameProgress';
+import { Question } from '@components/Question';
+import { AnswersList } from '@components/AnswersList';
+import { ShowInfo } from '@components/ShowInfo';
 
 const getAnswerShow = (shows: IShow[]): IShow => {
   const [answerShow] = shows.filter((show) => show.isAnswer);
   return answerShow;
 };
 
-const PlayPage: React.FC<IPlayPage> = ({
+export const PlayPage: React.FC<IPlayPage> = ({
   score,
   roundsData,
   handleClickToAnswer,
