+++ src/sass/pages/_start.scss
@@ -1,4 +1,5 @@
 .page_start {
+  padding: 0 2rem;
   text-align: center;
 
   @include flex($direction: column);
